disclosure-scraper/requirements.txt
requests==2.31.0
urllib3==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
colorlog==6.8.2
